Extracts randomize utils to a file
